Time: 1632 ms (5.01%), Space: 508.1 MB (5%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -7,9 +7,9 @@ public:
             for(int j = i + 1; j<n; j++){
                 set<long long> hashset;
                 for(int k = j+1; k<n; k++){
-                    int sum = nums[i] + nums[j];
+                    long long sum = nums[i] + nums[j];
                     sum += nums[k];
-                    long long fourth = target - (sum);
+                    long long fourth = (long long)target - (sum);
                     if(hashset.find(fourth) != hashset.end()){
                         vector<int> temp = {nums[i], nums[j], nums[k], (int) fourth};
                         sort(temp.begin(), temp.end());
